refactor(sequencer): Address review comments
