+++ apps/sequencer/src/http_handlers/mod.rs
@@ -1,1 +1,3 @@
-pub mod http_handlers;
+pub mod admin;
+pub mod data_feeds;
+pub mod registry;
